pages/requirements.txt
streamlit
pandas
yfinance
numpy
plotly
pandas_ta
tensorflow
keras
ta
scikit-learn
python-dateutil
